+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Render errors as JSON.
+  [jone]
 
 
 1.0a1 (2015-08-01)